Implement a basic QtWidgets UI. This only implements the "auto" mode, or "basic" mode as I now call it. The "Manual"/"Advanced" mode will be implemented later.
