+++ .gitignore
@@ -1,3 +1,5 @@
 /bin/
 /bin/*
 *.sky
+*qmake*
+Makefile
