+++ skymake.pro
@@ -1,0 +1,5 @@
+TARGET = skymake
+SOURCES += skymake.cpp
+QT += core gui widgets
+INCLUDEPATH += /usr/include/qt5 
+LIBS += -lQt5Core -lQt5Gui -lQt5Widgets
